Flattens destroys into the other command buffer commands (rationale follows)
You can't properly destroy entities with hierarchical relationships if the destroys
are not in order with the rest of the commands. It's tempting to think it works out
if you do the destroys first, and make parenting to a destroyed entity equivalent to
destroying an entity, but this doesn't actually work.

Consider parenting A to B, and then reparenting it to C, and then destroying B. A
should still exist after these operations!

Other systems may also care about order like this, so it's worth preserving it.

Also introduces a fuzzer specifically for command buffer encoding in an effort to track
down a bug that turned out to be elsewhere. I'm keeping the tests though, because they're
able to catch some other cases quickly that take a while with the other fuzzers.