Simplifies extension commands
They're no longer part of archetype change batches. In practice this
didn't offer much savings, and was a source of constant errors because
it's extremely difficult to write extensions in that form correctly
even with fuzz testing.

This was my biggest concern about this ECS design, and this new approach
resolves that concern.