Makes chunk size completely dynamic and removes page size default
The page size default is too small, and we don't actually want this to
vary. Most of the time we'll be a multiple of page size by virtue of being
a large power of two, and in the cases where we aren't it's because the page
size is way too large to be reasonable for this use case anyway.

This is also a step towards maybe being able to use smaller indices for chunk
indices, which removes memory overhead in chunks.