Modifies the view iterator to use get comp
This is technically slower and not what we want to actually merge w/ this PR, but
it will let us transition where we store components without having to sync two
different ways of retrieving them. We're going to completely replace this iterator
once the new location is set up anyway, and just use the buffer getters instead.