Better transform API, see long summary
The way transforms were handled previously was very clever, but turned
out to be 100x slower than the simpler approach!

The issue is that while it allowed for deferring the tranform sync and
then executing it in paralell on a thread pool, the transform sync is
actually very very fast, and queuing up this data then reading it back
into the cache again later is very very slow.

Even if a mild amount of work is repeated, it's easily faster to just
apply these changes immediately.

The real downside here is that you can't do a parallel update of the
ECS where a bunch of entities modify their own transforms at once. In
practice I suspect a lot of these updates will be so fast that doing them
in parallel doesn't make sense for the same reason though.

If we ever *do* need this ability, we could revive this algorithm, but I
think there are probably better simpler approaches. I almost want to say
you should just lock on the transforms since contention should be low
anyway, but then you get varying results based on order. You may want to
just somehow rule out that cross thread interaction, e.g. only do this
for roots somehow etc.

Another option is to bring back queuing it up, but handle it producer
consumer style, so even though there's overhead the bg threads get a
head start. Not worth implementing this unless we find real cases where
we want it though.