Adds Node.Tree which tracks the first child node
This allows us to assign order to nodes even when they have no parent,
by essentially creating an implicit null root node. We don't want to have
an actual root node since there's no way to prevent you from removing it
as a parent, this unfortuantely results in some special cases, but the main
annoyance is that both approaches result in more complexity around initializaiton.

IMO this is well worth it. It's expected to be able order transform nodes, it's weird
that there's no API for this right now but you can kinda pull it off until suddenly
you can't when you have no parent.

I'm going to follow this up with som eimproved tests, and then I'm going to
add immediate functions and extensions for assigning next/prev sib instead of
parent so that you can, when desired, insert objects directly into a speicfic order.

This is motivated by the fact that I'm currently leaning on the node graph a lot in
G2 for things like hand order. It's possible I'll walk this back, but this feature
seems generally reasonable regardless.