Removes the `order` field from ext.Transform
The intention of this field was to allow (for example) "2.5D" renderers
to create decals by forcing a sort on matching depths when using
painter's algorithm.

In practice, this doesn't work, because if the decal is offcenter the
depth won't match.

There are many ways to implement decals, and each are too renderer
specific to belong on transform. What I ended up opting for in my
current "2.5D" project is a `Decal` component that has a `target` entity,
and its z value (for the purposes of sorting) is overriden by that
of the target entity. There is also an order field on decal to force it
to sort above/below the target or any other decals.

This idea is simple enough, and also game specific enough, that I don't
think it really belongs in an extension here so I've left it off.

I've kept layer around because this is a generally useful feature that
I've also seen in other engines. I'm currently using it to render the
background at the focal length without clipping distant objects, as I
don't currently want parallax on the background.